backend/requirements.txt
bcrypt==4.1.2
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
Flask==3.0.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
greenlet==3.3.2
gunicorn==22.0.0

idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==26.0
PyJWT==2.11.0
python-dotenv==1.0.1
razorpay==2.0.0
requests==2.32.5
setuptools==68.0.0
SQLAlchemy==2.0.29
typing_extensions==4.15.0
urllib3==2.6.3
waitress==3.0.2
Werkzeug==3.0.3
wheel==0.46.3

google-auth

